requirements.txt
APScheduler==3.3.0
click==6.6
colorama==0.3.7
decorator==4.0.10
docopt==0.6.2
Flask==0.11.1
Flask-Cors==3.0.2
Flask-JWT==0.3.2
Flask-PyMongo==0.4.1
Flask-SocketIO==2.7.2
funcsigs==1.0.2
futures==3.0.5
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pipreqs==0.4.4
psutil==5.0.0
PyJWT==1.4.2
pymongo==2.9
python-engineio==1.0.3
python-socketio==1.6.0
pytz==2013.7
requests==2.10.0
six==1.10.0
tornado==2.4.1
tzlocal==1.3
validators==0.11.0
virtualenv==15.0.3
virtualenvwrapper-win==1.2.1
Werkzeug==0.11.10
yarg==0.1.9
eventlet==0.16.1
